Mouse Correction, Move Type Added
Move type makes every thing a lot easier. I noticed Dawson's code was working around mine to convert to points and when I found that built in data type realized, we should be using it. I incorporated this into a data type called Move that has some built in functions for making integers of the point coordinates for ease of implementation.

This drastically improves code efficiency and also ties in perfectly with the history function for keeping track of moves and captures. The app is no more seamlessly integrated and shares functionality.